requirements.txt
#pip install fastapi sqlalchemy psycopg2-binary uvicorn

# FastAPI & ASGI Server
fastapi
uvicorn

# PostgreSQL + ORM
sqlalchemy
psycopg2-binary

# Pydantic for validation
pydantic

# gRPC
grpcio
grpcio-tools

# SOAP
spyne
lxml  # pour validation XML avec SOAP
zeep

# Optional: to make environment explicit
python-dotenv  # si tu charges des variables .env
